pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "esn-pulse"
version = "1.0.0"
description = "ESN Network Data Scraper & Analytics Platform"
authors = [
    {name = "ESN PULSE Team", email = "esnpulse@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["esn", "scraping", "analytics", "data", "erasmus"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pydantic[email]>=2.0.0",
    "aiohttp>=3.8.0,<4.0.0",
    "beautifulsoup4>=4.12.0",
    "asyncpg>=0.28.0,<1.0.0",
    "celery>=5.3.0,<6.0.0",
    "redis>=5.0.0,<6.0.0",
    "click>=8.1.0",
    "python-slugify>=8.0.0",
    "python-dateutil>=2.8.0",
    "structlog>=23.1.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]

docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0"
]

monitoring = [
    "prometheus-client>=0.17.0"
]

[project.urls]
Homepage = "https://github.com/erdemgunal/esn-pulse"
Documentation = "https://esn-pulse.readthedocs.io/"
Repository = "https://github.com/erdemgunal/esn-pulse.git"
Issues = "https://github.com/erdemgunal/esn-pulse/issues"

[project.scripts]
esn-pulse = "src.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests"
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "redis.*",
    "asyncpg.*",
    "bs4.*"
]
ignore_missing_imports = true 